+++ src/backup/restic/eBooks.rec.ps1
@@ -1,8 +1,22 @@
-#----------------------------------------------------------------------
-#region Sauvegarde du dossier development sur Storj
-Write-Host ("-" * 44)
-Write-Host "| Recuperation du dossier eBooks sur Storj |"
-Write-Host ("-" * 44)
+Start-Transcript `
+    -Path D:\automation\log\eBooks.restic.log `
+    -Append -IncludeInvocationHeader
+
+    #----------------------------------------------------------------------
+#region Recuperation du dossier eBooks sur Storj
+Write-Host ("-" * 38)
+Write-Host "| Restoring eBooks folder from Storj |"
+Write-Host ("-" * 38)
+
+if (-not(Test-Path -Path Env:\RESTIC_REPOSITORY)) {
+    Write-Host "eBooks.rec.ps1 -- RESTIC_REPOSITORY empty"
+    Exit 1
+}
+
+if (-not(Test-Path -Path Env:\RESTIC_PASSWORD)) {
+    Write-Host "eBooks.rec.ps1 -- RESTIC_PASSWORD empty"
+    Exit 1
+}
 
 $dst = "D:\francois\eBooks"
 if (Test-Path -Path $dst -PathType Container) {
@@ -15,5 +29,15 @@ $params = @(
     '--target=D:\'
 )
 restic restore latest @params
-Move-Item -Path "D:\D\rancois" -Destination "D:\" -Force
-Remove-Item -Path "D:\D"
+if ($?) {
+    Write-Host "eBooks.rec.ps1 -- Data restoration successful"
+
+    Write-Host "eBooks.rec.ps1 -- Moving folder to original location"
+    Move-Item -Path "D:\D\francois" -Destination "D:\" -Force
+    Remove-Item -Path "D:\D"
+} else {
+    Write-Host "eBooks.rec.ps1 -- Data restoration failed"
+    Exit 1
+}
+
+Stop-Transcript
